fix: cleanup log event entry by moving processing code into a view function
